requirements.txt
fastapi==0.115.11
psycopg2-binary==2.9.10
uvicorn==0.34.0
dotenv==0.9.9